+++ src/main/java/com/pingplusplus/model/RoyaltyDataResult.java
@@ -1,0 +1,42 @@
+package com.pingplusplus.model;
+
+import com.pingplusplus.net.APIResource;
+
+public class RoyaltyDataResult extends APIResource {
+    String app;
+    Boolean livemode;
+    String  order;
+    Boolean succeeded;
+
+    public String getApp() {
+        return app;
+    }
+
+    public void setApp(String app) {
+        this.app = app;
+    }
+
+    public Boolean getLivemode() {
+        return livemode;
+    }
+
+    public void setLivemode(Boolean livemode) {
+        this.livemode = livemode;
+    }
+
+    public String getOrder() {
+        return order;
+    }
+
+    public void setOrder(String order) {
+        this.order = order;
+    }
+
+    public Boolean getSucceeded() {
+        return succeeded;
+    }
+
+    public void setSucceeded(Boolean succeeded) {
+        this.succeeded = succeeded;
+    }
+}